breseq + rebreseq scripts

+++ bash/rebreseq.sh
@@ -1,37 +1,61 @@
 #!/bin/bash
 
-bin_rebreseq_scripts="/N/dc2/projects/muri2/Task2/PoolPopSeq/bin/rebreseq_scripts"
-data_rebreseq="/N/dc2/projects/muri2/Task2/PoolPopSeq/data/rebreseq"
-data_rebreseq_out="/N/dc2/projects/muri2/Task2/PoolPopSeq/data/rebreseq_out"
-data_rebreseq_err="/N/dc2/projects/muri2/Task2/PoolPopSeq/data/rebreseq_err"
+bash_rebreseq_scripts="/N/dc2/projects/muri2/Task2/Phylo_Evol_Timeseries/bash/rebreseq_scripts"
+data_rebreseq="/N/dc2/projects/muri2/Task2/Phylo_Evol_Timeseries/data/rebreseq"
+data_rebreseq_out="/N/dc2/projects/muri2/Task2/Phylo_Evol_Timeseries/data/rebreseq_out"
+data_rebreseq_err="/N/dc2/projects/muri2/Task2/Phylo_Evol_Timeseries/data/rebreseq_err"
 
-mkdir -p $bin_rebreseq_scripts
+mkdir -p $bash_rebreseq_scripts
 mkdir -p $data_rebreseq
 mkdir -p $data_rebreseq_out
 mkdir -p $data_rebreseq_err
 
 ref=/N/dc2/projects/muri2/Task2/PoolPopSeq/data/reference_assemblies_task2/Bacillus_subtilis_NCIB_3610/GCA_002055965.1_ASM205596v1_genomic.gbff
 
-for line_path in "/N/dc2/projects/muri2/Task2/PoolPopSeq/data/reads_clean_trimmomatic/"*;
+declare -a strains=("B")
+declare -a treats=("0")
+declare -a reps=("2")
+
+declare -a times=("200")
+
+#declare -a times=("100" "300" "400" "500" "600" "700" "800" "900" "1000")
+
+
+declare -a samples=()
+
+for treat in "${treats[@]}"
 do
-  line="$(echo "$line_path" | cut -d "/" -f10-10)"
-  pop="$(echo "$line_path" | cut -d "/" -f10-10 | cut -d "-" -f1-1)"
-  pop_gd="/N/dc2/projects/muri2/Task2/PoolPopSeq/data/breseq_jc/merged/${pop}.gd"
-  bash_out="${bin_rebreseq_scripts}/${line}_rebreseq.sh"
-  taxon="$(echo "$line_path" | grep -Po ".(?=.{5}$)")"
-  if [ -f $bash_out ]; then
-    rm $bash_out
-  fi
-  taxon="$(echo "$line_path" | grep -Po ".(?=.{5}$)")"
-  if [[ $taxon == "B" ]] || [[ $taxon == "S" ]]; then
-    reads="${line_path}/"*_clean_paired.fastq.gz
-    OUT_rebreseq="${data_rebreseq}/${line}"
+  for strain in "${strains[@]}"
+  do
+    for rep in "${reps[@]}"
+    do
+      for time in "${times[@]}"
+      do
+        samples+=("${treat}${strain}${rep}_${time}")
+      done
+    done
+  done
+done
+
+
+
+for sample in "${samples[@]}"
+do
+  reads="/N/dc2/projects/muri2/Task2/Phylo_Evol_Timeseries/data/reads_clean_cutadapt/"*"${sample}"*"_clean.fastq.gz"
+  if (( ${#reads[@]} )); then
+    pop="$(echo "$sample" | cut -d "_" -f1-1)"
+    pop_gd="/N/dc2/projects/muri2/Task2/Phylo_Evol_Timeseries/data/breseq_jc/merged/${pop}.gd"
+    bash_out="${bash_rebreseq_scripts}/${sample}_rebreseq.sh"
+    if [ -f $bash_out ]; then
+      rm $bash_out
+    fi
+    OUT_rebreseq="${data_rebreseq}/${sample}"
     mkdir -p $OUT_rebreseq
-    OUT_rebreseq_out="${data_rebreseq_out}/${line}.out"
+    OUT_rebreseq_out="${data_rebreseq_out}/${sample}.out"
     if [ -f $OUT_rebreseq_out ]; then
       rm $OUT_rebreseq_out
     fi
-    OUT_rebreseq_err="${data_rebreseq_err}/${line}.err"
+    OUT_rebreseq_err="${data_rebreseq_err}/${sample}.err"
     if [ -f $OUT_rebreseq_err ]; then
       rm $OUT_rebreseq_err
     fi
@@ -49,7 +73,10 @@ do
     echo '' >> $bash_out
 
     qsub $bash_out
+
+
   else
     continue
   fi
+
 done
